api/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
mysql-connector-python==8.0.32
PyJWT==2.6.0
Werkzeug==2.2.3
python-dotenv==1.0.0
plaid-python==13.0.0
cryptography==40.0.2
